requirements.txt
# Core backend
fastapi
uvicorn
sqlalchemy
psycopg2-binary


langchain
langchain-community
langchain-openai
openai
Faker

pandas
# Optional: logging, pydantic (included via FastAPI), etc.
